Creating styled components files
