+++ src/components/ProfileScreen/index.jsx
@@ -1,23 +1,6 @@
-import styled from 'styled-components';
 import Header from '../MainHeaderSection';
 import Footer from '../MainFooterSection';
-
-const Main = styled.div`
-  height: 82vh;
-  background-color: #b1cdd8;
-`;
-
-const Container = styled.div`
-  display: flex;
-  justify-content: center;
-  border-radius: 20px;
-  background-color: #0d4d5d;
-`;
-
-const ProfileImg = styled.img`
-  height: 320px;
-  border-radius: 50%;
-`;
+import { Main, Container, ProfileImg } from '../Style/Profile';
 
 export default function ProfileScreen() {
   return (
