+++ src/components/Style/MainFooter/index.js
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+export const Footer = styled.footer`
+  display: flex;
+  flex-direction: column;
+  font-family: 'Montserrat', sans-serif;
+  align-items: center;
+  justify-content: center;
+  background-color: #ffffff;
+  min-height: 10vh;
+`;
+
+export const Logo = styled.img`
+  width: 35px;
+`;
+
+export const Label = styled.h1`
+  font-size: 15px;
+`;
