+++ src/components/Style/MainHeader/index.js
@@ -1,0 +1,28 @@
+import styled from 'styled-components';
+
+export const Header = styled.header`
+  height: 8vh;
+  background-color: #0d4d5d;
+  font-family: 'Montserrat', sans-serif;
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+`;
+
+export const MainTitle = styled.h1`
+  color: #ffffff;
+  font-size: 40px;
+`;
+
+export const Input = styled.input`
+  font-size: 18px;
+  border: none;
+  font-family: 'Montserrat', sans-serif;
+  padding: 5px 400px 5px 20px;
+  border-radius: 20px;
+`;
+
+export const ProfileImg = styled.img`
+  border-radius: 50%;
+  height: 50px;
+`;
