+++ src/components/Style/Profile/index.js
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const Main = styled.div`
+  height: 82vh;
+  background-color: #b1cdd8;
+`;
+
+export const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  border-radius: 20px;
+  background-color: #0d4d5d;
+`;
+
+export const ProfileImg = styled.img`
+  height: 320px;
+  border-radius: 50%;
+`;
